+++ .github/workflows/gnoland.yml
@@ -29,6 +29,7 @@ jobs:
           - gnokey
           - gnoweb
           - gnofaucet
+          - gnotxsync 
     runs-on: ubuntu-latest
     timeout-minutes: 5
     steps:
