+++ Dockerfile
@@ -10,7 +10,7 @@ RUN         go build -o ./build/gnokey    ./gno.land/cmd/gnokey
 RUN         go build -o ./build/gnofaucet ./gno.land/cmd/gnofaucet
 RUN         go build -o ./build/gnoweb    ./gno.land/cmd/gnoweb
 RUN         go build -o ./build/gno       ./gnovm/cmd/gno
-RUN         go build -o ./build/tm2txsync  ./tm2/cmd/tm2txsync
+RUN         go build -o ./build/gnotxsync  ./gno.land/cmd/gnotxsync
 RUN         ls -la ./build
 ADD         . /opt/gno/src/
 RUN         rm -rf /opt/gno/src/.git
@@ -42,9 +42,9 @@ COPY        --from=build /opt/build/build/gnofaucet /opt/gno/bin/
 ENTRYPOINT  ["gnofaucet"]
 EXPOSE      5050
 
-FROM        runtime-tls AS tm2txsync-slim
-COPY        --from=build /opt/build/build/tm2txsync /opt/gno/bin/
-ENTRYPOINT  ["tm2txsync"]
+FROM        runtime-tls AS gnotxsync-slim
+COPY        --from=build /opt/build/build/gnotxsync /opt/gno/bin/
+ENTRYPOINT  ["gnotxsync"]
 
 FROM        runtime-tls AS gnoweb-slim
 COPY        --from=build /opt/build/build/gnoweb /opt/gno/bin/
