+++ tm2/Makefile
@@ -7,10 +7,8 @@ help:
 rundep=go run -modfile ../misc/devdeps/go.mod
 
 .PHONY: build
-build: _build.tm2txsync
 _build.tools: _build.aminoscan _build.goscan _build.logjack _build.iaviewer
 
-_build.tm2txsync:;  go build -o build/tm2txsync  ./cmd/tm2txsync
 _build.aminoscan:;  go build -o build/aminoscan  ./pkg/amino/cmd/aminoscan
 _build.goscan:;     go build -o build/goscan     ./pkg/amino/cmd/goscan
 _build.logjack:;    go build -o build/logjack    ./pkg/autofile/cmd
@@ -18,7 +16,7 @@ _build.iaviewer:;   go build -o build/iaviewer   ./pkg/iavl/cmd/iaviewer
 
 .PHONY: install
 install:
-	go install ./cmd/...
+#	go install ./cmd/...
 
 .PHONY: clean
 clean:
@@ -36,7 +34,7 @@ lint:
 ########################################
 # Test suite
 .PHONY: test
-test: _test.pkg.amino _test.pkg.bft _test.pkg.others _test.flappy _test.cmds
+test: _test.pkg.amino _test.pkg.bft _test.pkg.others _test.flappy
 
 _test.flappy:
 	# flappy tests should work "sometimes" (at least once).
@@ -48,4 +46,3 @@ GOTEST_FLAGS ?= -v -p 1 -timeout=30m
 _test.pkg.others:;  go test $(GOTEST_FLAGS) `go list ./pkg/... | grep -v pkg/amino/ | grep -v pkg/bft/`
 _test.pkg.amino:;   go test $(GOTEST_FLAGS) ./pkg/amino/...
 _test.pkg.bft:;	    go test $(GOTEST_FLAGS) ./pkg/bft/...
-_test.cmds:;        go test $(GOTEST_FLAGS) ./cmd/...
